+++ selenium_driver_updater/consoleUpdater.py
@@ -12,16 +12,25 @@ class ConsoleUpdater():
     @staticmethod
     def parse_command_line():
         "Function for parsing arguments that were specified in command line"
+
+        description_text = (
+        "Download or update your selenium driver binaries and their browsers automatically with this package.\n\n"
+        "Available drivers are: (you can use them with -d or --driver_name command)\n"
+        "chromedriver, edgedriver, geckodriver, phantomjs, operadriver, safaridriver\r\n\r\n"
+        "Available OSes are: (you can use them with --system_name command)\n"
+        "win64, win32, linux64, linux32, mac64, mac64_m1 (stands for mac with arm-based architecture), arm64"
+        )
         parser = argparse.ArgumentParser(
-            description='Download or update your selenium driver binaries and their browsers automatically with this package.',
+            description=description_text, formatter_class=argparse.RawTextHelpFormatter
         )
         parser.add_argument(
         "--driver_name",
         "-d",
+        type=ConsoleUpdater.comma_separated_string,
         action="store",
         dest="driver_name",
-        metavar="F",
-        help="Specified driver name/names which will be downloaded or updated",
+        metavar="DRIVER_NAME",
+        help="Specified driver name/names which will be downloaded or updated, if you want to specify multiple drivers, use commas",
         default='',
         required=True,
         )
@@ -30,7 +39,7 @@ class ConsoleUpdater():
         "-p",
         action="store",
         dest="path",
-        metavar="F",
+        metavar="DIR",
         help="Specified path which will used for downloading or updating Selenium driver binary. Must be folder path",
         default='',
         )
@@ -39,7 +48,7 @@ class ConsoleUpdater():
         "-upg",
         action="store",
         dest="upgrade",
-        metavar="F",
+        metavar="BOOLEAN",
         help="If true, it will overwrite existing driver in the folder",
         default=False,
         )
@@ -48,7 +57,7 @@ class ConsoleUpdater():
         "-ch",
         action="store",
         dest="chmod",
-        metavar="F",
+        metavar="BOOLEAN",
         help="If true, it will make driver binary executable",
         default=True,
         )
@@ -57,7 +66,7 @@ class ConsoleUpdater():
         "-cdr",
         action="store",
         dest="check_driver_is_up_to_date",
-        metavar="F",
+        metavar="BOOLEAN",
         help="If true, it will check driver version before and after upgrade",
         default=True,
         )
@@ -66,15 +75,16 @@ class ConsoleUpdater():
         "-im",
         action="store",
         dest="info_messages",
-        metavar="F",
+        metavar="BOOLEAN",
         help="If false, it will disable all info messages",
         default=True,
         )
         parser.add_argument(
         "--filename",
+        type=ConsoleUpdater.comma_separated_string,
         action="store",
         dest="filename",
-        metavar="F",
+        metavar="FILENAME",
         help="Specific name for driver. If given, it will replace name for driver",
         default='',
         )
@@ -83,7 +93,7 @@ class ConsoleUpdater():
         "-cb",
         action="store",
         dest="check_browser_is_up_to_date",
-        metavar="F",
+        metavar="BOOLEAN",
         help="If true, it will check browser version before specific driver update or upgrade",
         default=False,
         )
@@ -91,12 +101,19 @@ class ConsoleUpdater():
         "--system_name",
         action="store",
         dest="system_name",
-        metavar="F",
+        metavar="SYSTEM_NAME",
         help="Specific OS for driver",
         default='',
         )
         parser.add_argument("--version", action="version", version=str(setting["Program"]["version"]))
         return parser.parse_args()
+    
+    @staticmethod
+    def comma_separated_string(value):
+        """Convert a comma-separated string into a list or return as a string if no comma."""
+        if ',' in value:
+            return value.split(',')
+        return value
 
     @staticmethod
     def install():
@@ -104,5 +121,9 @@ class ConsoleUpdater():
 
         args = ConsoleUpdater.parse_command_line()
         kwargs = vars(args)
+        
+        if kwargs['filename']:
+            if isinstance(kwargs['driver_name'], list) and isinstance(kwargs['filename'], str):
+                kwargs['filename'] = [kwargs['filename']]
 
         DriverUpdater.install(**kwargs)